Splits dynamic reports file into clases, replaces administrative units and dependencies filters and columns in datagrid
